% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_path.R
\name{network_path}
\alias{network_path}
\title{Create a Path to the QSP Network Drive}
\usage{
network_path(
  windows_path = fs::path("R:"),
  mac_path = fs::path("/Volumes", "Research")
)
}
\arguments{
\item{windows_path}{The path to use on Windows machines.}

\item{mac_path}{The path to use on Macs.}
}
\value{
a fs::path() object
}
\description{
This function creates a path to the QSP directory on the network drive. It
will automatically determine if the local machine runs Windows or MacOS, and
set the appropriate start of the network path.  This is not yet set up for
Linux machines.
}
\details{
This function will create the path object.  To use it, you must either be on
the Brown network or VPN into it.
}
\examples{
my_path <- network_path()
my_project <- fs::path(network_path(), "Projects", "My_Project")

}
